Use string instead of `{}` for spawning rc.exe
Seemingly in some recent version of PowerShell, something about how this command was being interpreted changed and it was sending a broken command line string when spawning rc.exe. Instead of

    /I "C:\some\path\to\include"

it was sending it as

    /I "\"C:\some\path\to\include\""

instead.

Instead of figuring out the proper root cause and/or reporting a possible bug, just switch to using cmd /C with the command as a string. This is how the non-rc.exe compilers are spawned anyway so it's not a crazy change.